requirements.txt
langchain==0.3.25
langchain_community==0.3.24
langchain_openai==0.3.18
python-dotenv==1.1.0
Requests==2.32.3
